requirements.txt
openai-whisper
streamlit
googletrans==4.0.0-rc1
